Add files via upload
This is v1.0.0 of an anomaly detection algorithm for transient objects. It utilizes two machine learning techniques to identify abnormal transient object light curves: Gaussian Mixture Models and Isolation Forests. This algorithm attaches anomaly scores for both techniques and identifies them as an anomaly if a certain threshold is passed (see Phillips et. al, in progress, for more details).